run the deduplication job on an interval
